requirements.txt
faker
kafka-python
simplejson
python-dateutil